refactor(tests): #58 Tier 1.2 — split ECB calendar scaffold tests
tests/test_ecb_calendar_api_scaffold.py (1,026 LOC, 32 tests) →
tests/calendar/ecb/{registry_parser,projector,fetch,mp_decision,service_op}.py.

  test_registry_parser.py   9 tests — INDICATOR_REGISTRY shape +
                                       parse_observation behaviour — 183 LOC
  test_projector.py         3 tests — store_raw / project_events /
                                       v_calendar_item — 103 LOC
  test_fetch.py            10 tests — fetch_ecb_calendar dry-run /
                                       persistence / step-change collapsing
                                       — 396 LOC
  test_mp_decision.py       6 tests — MP-decision ↔ SDMX rate
                                       reconciliation (P3a follow-up)
                                       — 448 LOC
  test_service_op.py        4 tests — calendar_econ_fetch_ecb /
                                       calendar_econ_schedule_ecb wiring
                                       — 80 LOC

Each split file mirrors the original's full import surface (verbatim) and
inlines `_dfr_obs` / `_mro_obs` / `_mlf_obs` row builders, the
`_FakeECBClient` stub, and `_seed_mp_decision_row` helper into whichever
file needs them. Test bodies are byte-equivalent.

Verification: 32 tests pass on branch (same as master baseline).
